Typecast correctly - compile breaks on 64 bit platforms otherwise.

+++ src/sys/dev/iwa/if_iwa_fw_util.c
@@ -420,7 +420,7 @@ iwa_mvm_send_cmd_status(struct iwa_softc *sc,
 		    "%s: response len %d != expected %d len\n",
 		    __func__,
 		    resp_len,
-		    sizeof(*resp));
+		    (int) sizeof(*resp));
 		error = EIO;
 		goto out_free_resp;
 	}
